refactor: Improve calculation of mean and variance in ChromosomeProcessor

+++ src/hotspot3/stats.py
@@ -11,8 +11,8 @@ from typing import Tuple
 
 # Calculate p-values and FDR
 def p_and_r_from_mean_and_var(mean: np.ndarray, var: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
-    r = np.asarray(mean ** 2 / (var - mean), dtype=np.float16)
-    p = np.asarray(1 - mean / var, dtype=np.float16)
+    r = np.asarray(mean ** 2 / (var - mean), dtype=np.float32)
+    p = np.asarray(1 - mean / var, dtype=np.float32)
     return p, r
